+++ udata-tabular-preview/config.js
@@ -16,5 +16,5 @@ function _meta(name) {
     return _attr(`meta[name=${name}]`, "content");
 }
 
-export const tabular_csvapi_url = _meta("tabular-csvapi-url");
+export const tabular_api_url = _meta("tabular-api-url");
 export const tabular_page_size = _meta("tabular-page-size");