+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## Current (in progress)
+## 4.0.0 (2024-03-22)
 
+- Connect preview with hydra and remove csvapi + styles for preview [#32](https://github.com/opendatateam/udata-tabular-preview/pull/33)
+ Current (in progress)
 - Use pagination from package `@etalab/data.gouv.fr-components` [#34](https://github.com/opendatateam/udata-tabular-preview/pull/34)
 
 ## 3.1.0 (2024-01-23)